US9932 Add style for cell-vertical-bottom
DE4293 worked to get the Homepage Anywhere section of the Homepage to
display the associated image correctly. In that solution, an inline
styling was applied in the CMS. This body of work extracts that inline
styling and creats a new class in crds-styles to be applied to the cell
that contains that image. The inline styling in the CMS will be removed
and the new style will be applied to that element after this pull
request gets merged.